refactor: input요소 통일
#230
