+++ frontend/src/components/ui/milestone/MilestoneCategoryDropdown.tsx
@@ -1,6 +1,6 @@
 import { Dispatch, SetStateAction, useEffect, useMemo, useState } from 'react';
 
-import { Dropdown, DropdownOption, DropdownProps } from '@/components/common/formik/DropdownDdang';
+import Dropdown, { DropdownOption, DropdownProps } from '@/components/common/formik/Dropdown';
 import { MilestoneGroup } from '@/data/milestone';
 import { useMilestoneQuery } from '@/lib/hooks/useApi';
 import { Milestone, MilestoneCategory } from '@/types/milestone';
